requirements.txt
openai-agents>=0.0.7
openai>=1.50.0
google-genai>=1.0.0
fastapi>=0.115.0
uvicorn>=0.34.0
langfuse>=2.50.0,<3.0.0
modal>=0.73.0
gradio>=5.0.0
mcp>=1.3.0
python-dotenv>=1.0.0
httpx>=0.28.0
Pillow>=11.0.0